requirements.txt
langchain
langchain-openai
azure-core
openai
bs4
requests
tiktoken
python-dotenv
langchain-community
streamlit